refactor: extract verify + file_browser overlays (#357)
Twelfth slice. Two larger overlays bundled:

- draw_verify       (~190 lines) - identity verification with 1:1 vs group layouts
- draw_file_browser (~135 lines) - attachment file picker with size column

Co-located helpers moved with their consumer:
- format_safety_number  (private to verify) -> overlays/verify.rs
- format_file_size      (was pub(crate), now private)  -> overlays/file_browser.rs
- format_file_size_cases test moved to file_browser.rs's tests module

Cleaned up three orphan doc comments left over from the previous
help/search extraction.

ui/mod.rs: 3110 -> 2747 lines.
