+++ src/ui/mod.rs
@@ -17,6 +17,7 @@ use overlays::contacts::draw_contacts;
 use overlays::customize::draw_customize;
 use overlays::delete_confirm::draw_delete_confirm;
 use overlays::emoji_picker::draw_emoji_picker;
+use overlays::file_browser::draw_file_browser;
 use overlays::forward::draw_forward;
 use overlays::help::draw_help;
 use overlays::keybindings::draw_keybindings;
@@ -29,6 +30,7 @@ use overlays::search::draw_search;
 use overlays::settings::draw_settings;
 use overlays::settings_profile::draw_settings_profile_manager;
 use overlays::theme_picker::draw_theme_picker;
+use overlays::verify::draw_verify;
 use sidebar::draw_sidebar;
 use status_bar::draw_status_bar;
 
@@ -60,8 +62,8 @@ pub(super) const SETTINGS_POPUP_WIDTH: u16 = 50;
 pub(super) const SETTINGS_POPUP_HEIGHT: u16 = 25;
 pub(super) const CONTACTS_POPUP_WIDTH: u16 = 50;
 pub(super) const CONTACTS_MAX_VISIBLE: usize = 20;
-const FILE_BROWSER_POPUP_WIDTH: u16 = 60;
-const FILE_BROWSER_MAX_VISIBLE: usize = 20;
+pub(super) const FILE_BROWSER_POPUP_WIDTH: u16 = 60;
+pub(super) const FILE_BROWSER_MAX_VISIBLE: usize = 20;
 pub(super) const SEARCH_POPUP_WIDTH: u16 = 60;
 pub(super) const SEARCH_MAX_VISIBLE: usize = 15;
 const GROUP_MENU_POPUP_WIDTH: u16 = 40;
@@ -2062,360 +2064,8 @@ fn draw_autocomplete(frame: &mut Frame, app: &App, input_area: Rect) {
     frame.render_widget(popup, area);
 }
 
-fn draw_verify(frame: &mut Frame, app: &App, area: Rect) {
-    let theme = &app.theme;
-    let is_group = app
-        .active_conversation
-        .as_ref()
-        .and_then(|id| app.store.conversations.get(id))
-        .map(|c| c.is_group)
-        .unwrap_or(false);
-
-    let pref_height: u16 = if is_group { 18 } else { 14 };
-    let pref_width: u16 = 50;
-    let (popup_area, block) = centered_popup(
-        frame,
-        area,
-        pref_width,
-        pref_height,
-        " Verify Identity ",
-        theme,
-    );
-    let inner = popup_area.inner(ratatui::layout::Margin {
-        horizontal: 1,
-        vertical: 1,
-    });
-    let mut lines: Vec<Line> = Vec::new();
-
-    if app.verify.identities.is_empty() {
-        lines.push(Line::from(Span::styled(
-            "  No identity information available",
-            Style::default().fg(theme.fg_muted),
-        )));
-        lines.push(Line::from(""));
-        lines.push(Line::from(Span::styled(
-            "  Esc: close",
-            Style::default().fg(theme.fg_muted),
-        )));
-    } else if is_group {
-        // Group view: scrollable member list with trust badges
-        let member_rows = inner.height.saturating_sub(7) as usize; // reserve for safety number + footer
-        let scroll_offset = if app.verify.index >= member_rows {
-            app.verify.index - member_rows + 1
-        } else {
-            0
-        };
-        let end = (scroll_offset + member_rows).min(app.verify.identities.len());
-
-        for (i, identity) in app.verify.identities[scroll_offset..end].iter().enumerate() {
-            let actual_idx = scroll_offset + i;
-            let is_selected = actual_idx == app.verify.index;
-            let number = identity.number.as_deref().unwrap_or("unknown");
-            let name = app
-                .store
-                .contact_names
-                .get(number)
-                .cloned()
-                .unwrap_or_else(|| number.to_string());
-            let (badge, badge_color) = match identity.trust_level {
-                TrustLevel::TrustedVerified => ("\u{2713}", theme.accent),
-                TrustLevel::Untrusted => ("\u{26A0}", theme.warning),
-                TrustLevel::TrustedUnverified => ("\u{2500}", theme.fg_muted),
-            };
-            let prefix = if is_selected { "> " } else { "  " };
-            let style = if is_selected {
-                Style::default()
-                    .bg(theme.bg_selected)
-                    .fg(theme.fg)
-                    .add_modifier(Modifier::BOLD)
-            } else {
-                Style::default().fg(theme.fg)
-            };
-            let badge_style = if is_selected {
-                Style::default().bg(theme.bg_selected).fg(badge_color)
-            } else {
-                Style::default().fg(badge_color)
-            };
-            lines.push(Line::from(vec![
-                Span::styled(prefix.to_string(), style),
-                Span::styled(format!("{badge} "), badge_style),
-                Span::styled(name, style),
-            ]));
-        }
-
-        lines.push(Line::from(""));
-
-        // Show selected member's safety number
-        if let Some(identity) = app.verify.identities.get(app.verify.index) {
-            if !identity.safety_number.is_empty() {
-                lines.push(Line::from(Span::styled(
-                    "  Safety Number:",
-                    Style::default().fg(theme.fg_secondary),
-                )));
-                let sn = &identity.safety_number;
-                let formatted = format_safety_number(sn);
-                for row in formatted {
-                    lines.push(Line::from(Span::styled(
-                        format!("  {row}"),
-                        Style::default().fg(theme.fg),
-                    )));
-                }
-            } else {
-                lines.push(Line::from(Span::styled(
-                    "  Safety number not available",
-                    Style::default().fg(theme.fg_muted),
-                )));
-            }
-        }
-
-        lines.push(Line::from(""));
-        if app.verify.confirming {
-            lines.push(Line::from(Span::styled(
-                "  Compare safety numbers, then press v to confirm",
-                Style::default().fg(theme.warning),
-            )));
-        } else {
-            lines.push(Line::from(Span::styled(
-                "  j/k: navigate  v: verify  Esc: close",
-                Style::default().fg(theme.fg_muted),
-            )));
-        }
-    } else {
-        // 1:1 view: single identity with full details
-        let identity = &app.verify.identities[0];
-        let number = identity.number.as_deref().unwrap_or("unknown");
-        let name = app
-            .store
-            .contact_names
-            .get(number)
-            .cloned()
-            .unwrap_or_else(|| number.to_string());
-
-        lines.push(Line::from(Span::styled(
-            format!("  {} ({})", name, number),
-            Style::default().fg(theme.fg).add_modifier(Modifier::BOLD),
-        )));
-
-        let (trust_label, trust_color) = match identity.trust_level {
-            TrustLevel::TrustedVerified => ("\u{2713} Verified", theme.accent),
-            TrustLevel::Untrusted => ("\u{26A0} Untrusted", theme.warning),
-            TrustLevel::TrustedUnverified => ("\u{2500} Unverified", theme.fg_muted),
-        };
-        lines.push(Line::from(Span::styled(
-            format!("  Trust: {trust_label}"),
-            Style::default().fg(trust_color),
-        )));
-        lines.push(Line::from(""));
-
-        if !identity.safety_number.is_empty() {
-            lines.push(Line::from(Span::styled(
-                "  Safety Number:",
-                Style::default().fg(theme.fg_secondary),
-            )));
-            let formatted = format_safety_number(&identity.safety_number);
-            for row in formatted {
-                lines.push(Line::from(Span::styled(
-                    format!("  {row}"),
-                    Style::default().fg(theme.fg),
-                )));
-            }
-        } else {
-            lines.push(Line::from(Span::styled(
-                "  Safety number not available",
-                Style::default().fg(theme.fg_muted),
-            )));
-        }
-
-        lines.push(Line::from(""));
-        if !identity.fingerprint.is_empty() {
-            lines.push(Line::from(Span::styled(
-                format!("  Fingerprint: {}", identity.fingerprint),
-                Style::default().fg(theme.fg_muted),
-            )));
-            lines.push(Line::from(""));
-        }
-
-        if app.verify.confirming {
-            lines.push(Line::from(Span::styled(
-                "  Compare safety numbers, then press v to confirm",
-                Style::default().fg(theme.warning),
-            )));
-        } else {
-            lines.push(Line::from(Span::styled(
-                "  v: verify key  Esc: close",
-                Style::default().fg(theme.fg_muted),
-            )));
-        }
-    }
-
-    let popup = Paragraph::new(lines).block(block);
-    frame.render_widget(popup, popup_area);
-}
-
 /// Format a safety number string as groups of 5 digits, 6 per line.
-fn format_safety_number(sn: &str) -> Vec<String> {
-    let digits: String = sn.chars().filter(|c| c.is_ascii_digit()).collect();
-    let chunks: Vec<&str> = digits
-        .as_bytes()
-        .chunks(5)
-        .map(|chunk| std::str::from_utf8(chunk).unwrap_or(""))
-        .collect();
-    chunks.chunks(6).map(|row| row.join(" ")).collect()
-}
-
-/// Extract a snippet of text centered around the first match of `query`.
-/// Build spans with the matching portions highlighted.
-/// Uses character-level case-insensitive matching to avoid byte-boundary issues.
 /// Format a file size in human-readable form (B, K, M, G).
-pub(crate) fn format_file_size(bytes: u64) -> String {
-    if bytes < 1024 {
-        format!("{bytes}B")
-    } else if bytes < 1024 * 1024 {
-        format!("{}K", bytes / 1024)
-    } else if bytes < 1024 * 1024 * 1024 {
-        format!("{:.1}M", bytes as f64 / (1024.0 * 1024.0))
-    } else {
-        format!("{:.1}G", bytes as f64 / (1024.0 * 1024.0 * 1024.0))
-    }
-}
-
-fn draw_file_browser(frame: &mut Frame, app: &App, area: Rect) {
-    let theme = &app.theme;
-    let visible_count = FILE_BROWSER_MAX_VISIBLE.min(if app.file_picker.filtered.is_empty() {
-        1
-    } else {
-        app.file_picker.filtered.len()
-    });
-    let pref_height = visible_count as u16 + 5; // border + header + footer
-
-    let title = if app.file_picker.filter.is_empty() {
-        " Attach File ".to_string()
-    } else {
-        format!(" Attach File [{}] ", app.file_picker.filter)
-    };
-
-    let (popup_area, block) = centered_popup(
-        frame,
-        area,
-        FILE_BROWSER_POPUP_WIDTH,
-        pref_height,
-        &title,
-        theme,
-    );
-
-    let inner_height = popup_area.height.saturating_sub(2) as usize;
-    let header_lines = 1; // path header
-    let footer_lines = 2; // empty + key hints
-    let visible_rows = inner_height.saturating_sub(header_lines + footer_lines);
-    let inner_w = popup_area.width.saturating_sub(2) as usize;
-
-    let mut lines: Vec<Line> = Vec::new();
-
-    // Current path header
-    let dir_display = app.file_picker.dir.to_string_lossy();
-    let dir_truncated = truncate(&dir_display, inner_w.saturating_sub(2));
-    lines.push(Line::from(Span::styled(
-        format!("  {dir_truncated}"),
-        Style::default()
-            .fg(theme.accent)
-            .add_modifier(Modifier::BOLD),
-    )));
-
-    if let Some(ref err) = app.file_picker.error {
-        lines.push(Line::from(Span::styled(
-            format!("  {}", truncate(err, inner_w.saturating_sub(2))),
-            Style::default().fg(theme.error),
-        )));
-    } else if app.file_picker.filtered.is_empty() {
-        lines.push(Line::from(Span::styled(
-            "  Empty directory",
-            Style::default().fg(theme.fg_muted),
-        )));
-    } else {
-        // Scroll the list so the selected item is always visible
-        let scroll_offset = if app.file_picker.index >= visible_rows {
-            app.file_picker.index - visible_rows + 1
-        } else {
-            0
-        };
-
-        let end = (scroll_offset + visible_rows).min(app.file_picker.filtered.len());
-
-        for (i, &entry_idx) in app.file_picker.filtered[scroll_offset..end]
-            .iter()
-            .enumerate()
-        {
-            let actual_index = scroll_offset + i;
-            let is_selected = actual_index == app.file_picker.index;
-            let (ref name, is_dir, size) = app.file_picker.entries[entry_idx];
-
-            let size_str = if is_dir {
-                String::new()
-            } else {
-                format_file_size(size)
-            };
-
-            let display_name = if is_dir {
-                format!("{name}/")
-            } else {
-                name.clone()
-            };
-
-            // Leave room for size column
-            let size_col_width = 8;
-            let name_max = inner_w.saturating_sub(size_col_width + 4);
-            let display_name = truncate(&display_name, name_max);
-
-            let name_style = if is_selected {
-                if is_dir {
-                    Style::default()
-                        .bg(theme.bg_selected)
-                        .fg(theme.accent)
-                        .add_modifier(Modifier::BOLD)
-                } else {
-                    Style::default()
-                        .bg(theme.bg_selected)
-                        .fg(theme.fg)
-                        .add_modifier(Modifier::BOLD)
-                }
-            } else if is_dir {
-                Style::default().fg(theme.accent)
-            } else {
-                Style::default().fg(theme.fg)
-            };
-
-            let size_style = if is_selected {
-                Style::default().bg(theme.bg_selected).fg(theme.fg_muted)
-            } else {
-                Style::default().fg(theme.fg_muted)
-            };
-
-            // Pad name to align size column
-            let name_padded = format!("  {display_name:width$}", width = name_max);
-            let size_padded = format!("{size_str:>width$}  ", width = size_col_width);
-
-            lines.push(Line::from(vec![
-                Span::styled(name_padded, name_style),
-                Span::styled(size_padded, size_style),
-            ]));
-        }
-    }
-
-    // Pad to fill visible rows
-    while lines.len() < header_lines + visible_rows {
-        lines.push(Line::from(""));
-    }
-
-    lines.push(Line::from(""));
-    lines.push(Line::from(Span::styled(
-        "  j/k nav  Enter open/select  Backspace/- up  Esc cancel",
-        Style::default().fg(theme.fg_muted),
-    )));
-
-    let popup = Paragraph::new(lines).block(block);
-    frame.render_widget(popup, popup_area);
-}
-
 pub(crate) fn build_poll_display(
     poll: &PollData,
     votes: &[PollVote],
@@ -2759,19 +2409,6 @@ mod tests {
             summary.contains("0 votes"),
             "expected '0 votes' in: {summary}"
         );
-    }
-
-    // --- format_file_size ---
-
-    #[rstest]
-    #[case(0, "0B")]
-    #[case(512, "512B")]
-    #[case(1024, "1K")]
-    #[case(2048, "2K")]
-    #[case(1_048_576, "1.0M")]
-    #[case(1_073_741_824, "1.0G")]
-    fn format_file_size_cases(#[case] bytes: u64, #[case] expected: &str) {
-        assert_eq!(format_file_size(bytes), expected);
     }
 }
