+++ src/ui/overlays/mod.rs
@@ -11,6 +11,7 @@ pub(super) mod contacts;
 pub(super) mod customize;
 pub(super) mod delete_confirm;
 pub(super) mod emoji_picker;
+pub(super) mod file_browser;
 pub(super) mod forward;
 pub(super) mod help;
 pub(super) mod keybindings;
@@ -23,3 +24,4 @@ pub(super) mod search;
 pub(super) mod settings;
 pub(super) mod settings_profile;
 pub(super) mod theme_picker;
+pub(super) mod verify;
